python/requirements.txt
# Ultra-Face Recognition System - Requirements
# =============================================
# 工业级超极限人脸识别系统依赖

# ============================================================================
# 核心依赖
# ============================================================================
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ============================================================================
# 数据处理
# ============================================================================
opencv-python>=4.8.0
numpy>=1.24.0
albumentations>=1.3.0
Pillow>=9.5.0

# ============================================================================
# 检索
# ============================================================================
faiss-gpu>=1.7.4
# 或 faiss-cpu>=1.7.4 (无 GPU 支持)

# ============================================================================
# 服务部署
# ============================================================================
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
websockets>=11.0

# ============================================================================
# 工具
# ============================================================================
tqdm>=4.65.0
PyYAML>=6.0
requests>=2.31.0

# ============================================================================
# 开发工具
# ============================================================================
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0

# ============================================================================
# 可选依赖
# ============================================================================
# tensorboard>=2.13.0      # 训练可视化
# wandb>=0.15.0            # 实验跟踪
# onnx>=1.14.0             # ONNX 导出
# onnxruntime-gpu>=1.15.0  # ONNX 推理
